Accepting that CTX_get_options was signed was causing headaches
The flags in the python interpreter(s) are treated as positive numbers, but
if the high bit is set it comes back out of openssl as negative.  This is
pedantically incorrect but produces sanity-enhancing results.

(CTX_set_options is also altered similarly for consiswtency as it's the
same underlying function behind the macro)