(arigo, ltratt) Avoid allocating memory when ordereddicts have spare capacity.
Previously, dicts could yo-yo in size if someone continually added and removed
elements, but the total number of live elements remained more-or-less constant.
This patch changes things so that compaction of entries happens more often than
shrinking of memory. Put another way, we only shrink the size allocated to an
ordereddict when it is extremely sparsely populated. This reduces GC pressure
quite a bit.